bpf: Add lookup_and_delete_elem for BPF_MAP_STACK_TRACE
The stacktrace map can be easily full, which will lead to failure in obtaining
the stack. In addition to increasing the size of the map, another solution is
to delete the stack_id after looking it up from the user, so extend the existing
bpf_map_lookup_and_delete_elem() functionality to stacktrace map types.

Signed-off-by: Tao Chen <chen.dylane@linux.dev>